demo_script/requirements.txt
playwright>=1.47.0
html2text>=2024.2.26
requests>=2.31.0
python-dotenv>=1.0.1
openai>=1.51.0
pydantic>=2.8.2
fastapi>=0.104.0
uvicorn>=0.24.0
cachetools>=5.3.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0
